Rewrite *backandforthd in Python
This makes them much faster, which is important: alt-tab and friends
can't be run again until the relevant *backandforthd has noticed the
change and updated the on-disk state. For outputbackandforthd, this was
taking several hundred milliseconds (three jq invocations, each taking
around 70ms), which is *way* too long.

The non-daemon scripts are not rewritten, since the Python interpreter
takes quite a while to start up (though I haven't measured this for a
while, and it might be worth investigating), but that's not a thing we
care about for daemons.