Rewrite backandforth and i3-workspace-swap in Rust
This significantly reduces the latency of common windowing operations,
especially for i3-workspace-swap (was ~200ms, now ~30ms, presumably due
to heavy use of jq).

There is a minor behaviour change for backandforth: in the case of the
backandforthd bug occurring (the old workspace for an output no longer
being on that output), we no longer just exit immediately; instead, if
there are exactly two workspaces on the current output, we switch to the
other one. This hides the bug in the (very common) case of there being
two workspaces on an output, but provides a much-improved user
experience in almost all cases, the exception being if the correct
workspace to switch to would be created by the switch.