Fix i3status-rust patches
Previously, i3status-rust was being installed both system-wide and
through home-manager, but the patches were only being applied to the
system-wide version. We stop installing it system-wide and move the
overlay to the list of per-user overlays.

It would be nice to solve this in a way that makes more sense, but this
is the easiest way to make the patches actually take effect for the time
being.