Show the time in the prompt
This binding updates the prompt just before the command is executed (as
long as you press enter rather than e.g. ctrl-j), so the time shows when
the command was started, not when the prompt was first printed.

I'm extremely happy that this is so simple to do in Fish!