Add Lorri status to prompt
We have to use an internal Lorri command for this, which isn't ideal,
but hopefully it will work ok for the foreseeable future.

This slows down the prompt noticeably when in a directory with a
shell.nix, unfortunately. About 10ms of the extra time is from running
lorri, which increases by 20ms when adding in jq. It's not intolerable,
though, and it doesn't happen at all when not actively using Lorri.